Compiler -> OperationCompiler, Runner -> OperationRunner (#127)
