+++ modules/lsp/src/main/scala/playground/lsp/ServerBuilder.scala
@@ -12,7 +12,7 @@ import org.http4s.client.middleware.Logger
 import org.http4s.ember.client.EmberClientBuilder
 import org.http4s.headers.Authorization
 import playground.language.CommandResultReporter
-import playground.Runner
+import playground.OperationRunner
 import playground.TextDocumentManager
 import playground.std.StdlibRuntime
 import smithy4s.aws.AwsEnvironment
@@ -67,7 +67,7 @@ object ServerBuilder {
                         PluginResolver[F]
                           .resolveFromConfig(buildInfo.config)
                           .flatMap { plugins =>
-                            val runner = Runner
+                            val runner = OperationRunner
                               .forSchemaIndex[F](
                                 dsi,
                                 client,
